SQL projects and changed TestRuntime name
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C29781

+++ TOneV2/Code/TOneV2/SQL.TOneWhS_BE/TOneWhS_Analytic/User Defined Types/TrafficStatsType.sql
@@ -3,20 +3,22 @@
     [CustomerId]             INT             NULL,
     [SupplierId]             INT             NULL,
     [Attempts]               INT             NOT NULL,
-    [TotalDurationInSeconds] INT             NOT NULL,
+    [TotalDurationInSeconds] NUMERIC (20, 6) NOT NULL,
     [FirstCDRAttempt]        DATETIME        NOT NULL,
     [LastCDRAttempt]         DATETIME        NOT NULL,
     [SaleZoneID]             BIGINT          NULL,
     [SupplierZoneID]         BIGINT          NULL,
-    [PDDInSeconds]           INT             NOT NULL,
-    [MaxDurationInSeconds]   INT             NOT NULL,
+    [PDDInSeconds]           NUMERIC (20, 6) NOT NULL,
+    [MaxDurationInSeconds]   NUMERIC (20, 6) NOT NULL,
     [NumberOfCalls]          INT             NOT NULL,
     [PortOut]                NVARCHAR (50)   NULL,
     [PortIn]                 NVARCHAR (50)   NULL,
     [DeliveredAttempts]      INT             NOT NULL,
     [SuccessfulAttempts]     INT             NOT NULL,
-    [SumOfPGAD]              INT             NOT NULL,
+    [SumOfPGAD]              NUMERIC (20, 6) NOT NULL,
     [DeliveredNumberOfCalls] INT             NOT NULL,
     [CeiledDuration]         BIGINT          NOT NULL,
-    [UtilizationInSeconds]   NUMERIC (19, 5) NOT NULL);
+    [UtilizationInSeconds]   NUMERIC (20, 6) NOT NULL);
+
+
